Remove redundant type casts in AppendBodyTo where field type already matches
Agent-Logs-Url: https://github.com/0blu/vmangos-core/sessions/bd1f37bc-7de3-41aa-816e-52a49640fd2e

+++ src/game/Server/Packets/Item.cpp
@@ -152,7 +152,7 @@ void WorldPackets::Item::ReadItemOk::AppendBodyTo(ByteBuffer& buffer) const
 void WorldPackets::Item::ReadItemFailed::AppendBodyTo(ByteBuffer& buffer) const
 {
     buffer << itemGuid;
-    buffer << uint8(reason);
+    buffer << reason;
     buffer << itemGuid;
 }
 
@@ -172,19 +172,19 @@ void WorldPackets::Item::ListInventoryResponse::AppendBodyTo(ByteBuffer& buffer)
     if (items.empty())
     {
         buffer << uint8(0);
-        buffer << uint8(errorCode);
+        buffer << errorCode;
         return;
     }
     buffer << uint8(items.size());
     for (auto const& item : items)
     {
-        buffer << uint32(item.slot);
-        buffer << uint32(item.itemId);
-        buffer << uint32(item.displayInfoId);
-        buffer << uint32(item.currentCount);
-        buffer << uint32(item.price);
-        buffer << uint32(item.maxDurability);
-        buffer << uint32(item.buyCount);
+        buffer << item.slot;
+        buffer << item.itemId;
+        buffer << item.displayInfoId;
+        buffer << item.currentCount;
+        buffer << item.price;
+        buffer << item.maxDurability;
+        buffer << item.buyCount;
     }
 }
 
@@ -196,87 +196,87 @@ void WorldPackets::Item::ItemQuerySingleResponse::AppendBodyTo(ByteBuffer& buffe
         return;
     }
 
-    buffer << uint32(itemId);
-    buffer << uint32(itemClass);
-    buffer << uint32(subClass);
+    buffer << itemId;
+    buffer << itemClass;
+    buffer << subClass;
     buffer << name;
     buffer << uint8(0); // Name2
     buffer << uint8(0); // Name3
     buffer << uint8(0); // Name4
-    buffer << uint32(displayInfoId);
-    buffer << uint32(quality);
-    buffer << uint32(flags);
-    buffer << uint32(buyPrice);
-    buffer << uint32(sellPrice);
-    buffer << uint32(inventoryType);
-    buffer << uint32(allowableClass);
-    buffer << uint32(allowableRace);
-    buffer << uint32(itemLevel);
-    buffer << uint32(requiredLevel);
-    buffer << uint32(requiredSkill);
-    buffer << uint32(requiredSkillRank);
-    buffer << uint32(requiredSpell);
-    buffer << uint32(requiredHonorRank);
-    buffer << uint32(requiredCityRank);
+    buffer << displayInfoId;
+    buffer << quality;
+    buffer << flags;
+    buffer << buyPrice;
+    buffer << sellPrice;
+    buffer << inventoryType;
+    buffer << allowableClass;
+    buffer << allowableRace;
+    buffer << itemLevel;
+    buffer << requiredLevel;
+    buffer << requiredSkill;
+    buffer << requiredSkillRank;
+    buffer << requiredSpell;
+    buffer << requiredHonorRank;
+    buffer << requiredCityRank;
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_6_1
-    buffer << uint32(requiredReputationFaction);
-    buffer << uint32(requiredReputationRank);
+    buffer << requiredReputationFaction;
+    buffer << requiredReputationRank;
 #endif
-    buffer << int32(maxCount);
-    buffer << int32(stackable);
-    buffer << uint32(containerSlots);
+    buffer << maxCount;
+    buffer << stackable;
+    buffer << containerSlots;
     for (int i = 0; i < 10; ++i)
     {
-        buffer << uint32(itemStats[i].type);
-        buffer << int32(itemStats[i].value);
+        buffer << itemStats[i].type;
+        buffer << itemStats[i].value;
     }
     for (int i = 0; i < 5; ++i)
     {
-        buffer << float(damages[i].damageMin);
-        buffer << float(damages[i].damageMax);
-        buffer << uint32(damages[i].damageType);
+        buffer << damages[i].damageMin;
+        buffer << damages[i].damageMax;
+        buffer << damages[i].damageType;
     }
-    buffer << uint32(armor);
-    buffer << uint32(holyRes);
-    buffer << uint32(fireRes);
-    buffer << uint32(natureRes);
-    buffer << uint32(frostRes);
-    buffer << uint32(shadowRes);
-    buffer << uint32(arcaneRes);
-    buffer << uint32(delay);
-    buffer << uint32(ammoType);
+    buffer << armor;
+    buffer << holyRes;
+    buffer << fireRes;
+    buffer << natureRes;
+    buffer << frostRes;
+    buffer << shadowRes;
+    buffer << arcaneRes;
+    buffer << delay;
+    buffer << ammoType;
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_9_4
-    buffer << float(rangedModRange);
+    buffer << rangedModRange;
 #endif
     for (int i = 0; i < 5; ++i)
     {
-        buffer << uint32(spells[i].spellId);
-        buffer << uint32(spells[i].spellTrigger);
-        buffer << int32(spells[i].spellCharges);
-        buffer << uint32(spells[i].spellCooldown);
-        buffer << uint32(spells[i].spellCategory);
-        buffer << uint32(spells[i].spellCategoryCooldown);
+        buffer << spells[i].spellId;
+        buffer << spells[i].spellTrigger;
+        buffer << spells[i].spellCharges;
+        buffer << spells[i].spellCooldown;
+        buffer << spells[i].spellCategory;
+        buffer << spells[i].spellCategoryCooldown;
     }
-    buffer << uint32(bonding);
+    buffer << bonding;
     buffer << description;
-    buffer << uint32(pageText);
-    buffer << uint32(languageId);
-    buffer << uint32(pageMaterial);
-    buffer << uint32(startQuest);
-    buffer << uint32(lockId);
-    buffer << int32(material);
-    buffer << uint32(sheath);
-    buffer << int32(randomProperty);
-    buffer << uint32(block);
-    buffer << uint32(itemSet);
-    buffer << uint32(maxDurability);
+    buffer << pageText;
+    buffer << languageId;
+    buffer << pageMaterial;
+    buffer << startQuest;
+    buffer << lockId;
+    buffer << material;
+    buffer << sheath;
+    buffer << randomProperty;
+    buffer << block;
+    buffer << itemSet;
+    buffer << maxDurability;
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_6_1
-    buffer << uint32(area);
+    buffer << area;
 #endif
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_10_2
-    buffer << uint32(map);
+    buffer << map;
 #endif
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_8_4
-    buffer << uint32(bagFamily);
+    buffer << bagFamily;
 #endif
 }
